fix: fixed count-to-channel feature

+++ src/events/messageUpdate.js
@@ -1,7 +1,14 @@
-const messageUpdate = (client, oldMessage, newMessage) => {
+const messageUpdate = async (client, oldMessage, newMessage) => {
   if (client.features.countToChannel) {
+    // compromise that prevents one person from breaking the entire channel without it being possible for the bot to ruin anything
+
     if (newMessage.channel.name.startsWith('count-to-')) {
-      newMessage.delete();
+      const messageList = await newMessage.channel.messages.fetch({ limit: 1 });
+      const lastMessage = messageList.first();
+
+      if (newMessage === lastMessage) {
+        newMessage.delete();
+      }
     }
   }
 };
